+++ bench/index.js
@@ -1,7 +1,9 @@
 import interval from './suites/interval.js';
 import duration from './suites/duration.js';
 import date from './suites/date.js';
+import apply from './suites/apply.js';
 
 await interval();
 await duration();
 await date();
+await apply();